didnt see more regular link errors. fix
